+++ R/Multidiff_Plot.R
@@ -1545,6 +1545,8 @@ plot_multi_compare<-function(multi_compare_objects,plots_label=NULL, plot_title=
   for (i in 1:length(multi_compare_objects)){
 
     curr_df<-get(multi_compare_objects[i])
+    
+    if(is.function(curr_df)) stop(paste("multi_compare_objects must not be named the same as a existing function"))
 
 
 
@@ -1651,7 +1653,7 @@ plot_multi_compare<-function(multi_compare_objects,plots_label=NULL, plot_title=
                    plot.caption=ggplot2::element_text(hjust = 0))+
     ggplot2::ggtitle(plot_title)+
     ggplot2::guides(alpha="none")+
-    ggplot2::facet_wrap(~ factor(samp),ncol = ncol_facet)
+    ggplot2::facet_wrap(~ factor(samp,levels=unique(samp)),ncol = ncol_facet)
 
 
   if(note==TRUE) comparison_plot<-comparison_plot + ggplot2::labs(caption = note_text)